Refactor Player class and remove unused code
Now not needed and unused method The code for drawing the bounding box of the player has been removed as it is no longer needed.

Fix distance calculation in Game.js

In the Game.js file, fixed the calculation of distance between brazilPlayer and ball.possession. The previous calculation used the wrong exponentiation operator, causing incorrect results. Now, the correct multiplication operator is used.